refactor: move sender_allocation state to separated file
